updated with config tests
