--- backend/common/settings/settings_test.go
@@ -1,40 +1,0 @@
-package settings
-
-import (
-	"testing"
-
-	"github.com/google/go-cmp/cmp"
-)
-
-func TestConfigLoadChanged(t *testing.T) {
-	defaultConfig := setDefaults()
-	err := loadConfigWithDefaults("./testingConfig.yaml")
-	if err != nil {
-		t.Fatalf("error loading config file: %v", err)
-	}
-	// Use go-cmp to compare the two structs
-	if diff := cmp.Diff(defaultConfig, Config); diff == "" {
-		t.Errorf("No change when there should have been (-want +got):\n%s", diff)
-	}
-}
-
-func TestConfigLoadSpecificValues(t *testing.T) {
-	defaultConfig := setDefaults()
-	err := loadConfigWithDefaults("./testingConfig.yaml")
-	if err != nil {
-		t.Fatalf("error loading config file: %v", err)
-	}
-	testCases := []struct {
-		fieldName string
-		globalVal interface{}
-		newVal    interface{}
-	}{
-		{"Server.Database", Config.Server.Database, defaultConfig.Server.Database},
-	}
-
-	for _, tc := range testCases {
-		if tc.globalVal == tc.newVal {
-			t.Errorf("Differences should have been found:\nConfig.%s: %v \nSetConfig: %v \n", tc.fieldName, tc.globalVal, tc.newVal)
-		}
-	}
-}
